Improve Gemini JSON parsing reliability
