+++ vscode_controller_project3/README.md
@@ -1,256 +1,108 @@
-# AI 自動化開發控制器 Pro v2.0
+# AI 自動化開發控制器 Pro v3.2
 
-## 🚀 專案簡介
+AI 自動化開發控制器 Pro 是一套整合 Gemini AI 與 VS Code 自動化能力的桌面輔助工具。最新版本針對後端重新模組化，將高頻共享變量集中管理，避免跨檔案重複儲存，並補強偵錯與日誌資訊，協助開發者快速定位問題。
 
-AI 自動化開發控制器 Pro 是一個整合了 Gemini AI 與 VS Code 自動化控制的智能開發助手。本專案透過獨立桌面視窗提供友善的操作介面，能夠自動生成程式碼、管理專案檔案、監控執行狀態，並支援螢幕擷取功能。
+## 🔧 核心特色
 
-## ✨ 新版本特色 (v2.0)
-
-### 前端改進
-- **🎨 全新設計的明亮美觀界面**
-  - 採用現代化的漸層配色方案（藍紫色系）
-  - 卡片式布局設計，視覺層次分明
-  - 動畫效果提升互動體驗
-  - 響應式設計，支援不同螢幕尺寸
-
-- **📊 多標籤頁管理系統**
-  - 主控制台：核心功能操作
-  - 設定頁面：完整的配置管理
-  - 監控頁面：即時查看執行狀態
-  - 自動化頁面：工作流程設定
-
-- **🔄 即時狀態指示器**
-  - 系統狀態即時顯示
-  - 載入動畫與進度提示
-  - 操作結果視覺化反饋
-
-### 後端改進
-- **📦 模塊化架構設計**
-  - ConfigManager：配置文件管理
-  - GeminiAI：AI API 整合
-  - VSCodeController：VS Code 自動化控制
-  - ScreenCapture：螢幕擷取功能
-  - CodeProcessor：程式碼解析處理
-  - TerminalMonitor：Terminal 輸出監控
-  - ProcessManager：主流程管理
-
-- **📝 完整的日誌系統**
-  - 結構化日誌記錄
-  - 時間戳記與級別標記
-  - 錯誤追蹤與除錯支援
-
-- **🛡️ 增強的錯誤處理**
-  - 全面的異常捕獲
-  - 友善的錯誤訊息
-  - 失敗恢復機制
-
-### 新增功能
-- **📸 螢幕擷取功能 (使用 MSS)**
-  - 支援多螢幕擷取
-  - 特定視窗擷取（如 VS Code）
-  - 截圖自動儲存與管理
-  - 預覽功能整合
-
-- **🔍 VS Code 操作透明性**
-  - 詳細的操作步驟記錄
-  - 視窗狀態即時追蹤
-  - 文件開啟確認機制
-
-- **⚡ Terminal 監控優化**
-  - 執行超時控制（避免無限等待）
-  - 輸出緩衝區管理
-  - 錯誤輸出分離
-  - 退出碼檢測
-
-## 🔧 系統需求
-
-- Python 3.8 或更高版本
-- Visual Studio Code（需將 `code` 命令加入 PATH）
-- Windows 10/11、macOS 10.14+ 或 Linux（Ubuntu 20.04+）
-- Gemini API Key 或 Google Cloud 憑證
-
-## 📦 安裝指南
-
-### 步驟 1: 克隆專案
-```bash
-git clone https://github.com/your-repo/ai-controller-pro.git
-cd ai-controller-pro
-```
-
-### 步驟 2: 建立虛擬環境
-```bash
-# Windows
-python -m venv venv
-.\venv\Scripts\activate
-
-# macOS/Linux
-python3 -m venv venv
-source venv/bin/activate
-```
-
-### 步驟 3: 安裝依賴套件
-```bash
-pip install -r requirements.txt
-```
-
-### 步驟 4: 配置 API 連接
-
-#### 方式 A：使用 API Key
-1. 訪問 [Google AI Studio](https://aistudio.google.com/app/apikey)
-2. 獲取您的 Gemini API Key
-3. 在應用程式設定中輸入 Key
-
-#### 方式 B：使用 Google Cloud Auth
-```bash
-# 安裝 gcloud CLI 後執行
-gcloud auth application-default login
-```
-
-## 🎮 使用說明
-
-### 啟動應用程式
-```bash
-python app.py
-```
-
-### 基本操作流程
-
-1. **選擇專案資料夾**
-   - 點擊「瀏覽」按鈕選擇您的專案目錄
-
-2. **輸入 AI 指令**
-   - 在文字框中描述您需要的功能
-   - 支援安裝指令格式：`;;;pip install package;;;`
-   - 檔案名稱格式：`/*/filename.py/*/`
-
-3. **執行自動化**
-   - 點擊「執行 AI 自動化」按鈕
-   - 系統將自動：
-     - 呼叫 Gemini AI 生成程式碼
-     - 安裝所需套件
-     - 創建並儲存檔案
-     - 啟動 VS Code 並打開檔案
-     - 執行並監控程式
-
-4. **螢幕擷取**
-   - 點擊「擷取螢幕畫面」按鈕
-   - 自動擷取所有螢幕和 VS Code 視窗
-   - 截圖儲存在 `~/.ai_controller_v2/screenshots/`
+- **單一來源的資料模型**：所有 JSON 解析、專案結構與目錄設定皆集中於 `core.py`，確保 AI 回應的重點資訊不再分散。
+- **明確的服務分層**：
+  - `services.py` 專責設定檔存取與 Gemini API 呼叫。
+  - `automation.py` 提供 VS Code 操控、程式執行、螢幕擷取與整體流程管理。
+  - `app.py` 僅負責 Flask/Webview 入口與 API 路由，維持總檔案數量不超過 5。
+- **強化偵錯訊息**：流程關鍵步驟皆納入結構化日誌與錯誤說明，包含 AI 解析失敗時的可能原因與排錯建議。
+- **變量單點維護**：常用狀態（如螢幕截圖路徑、視窗標題、伺服器資訊）只在後端建立一次並全程沿用，避免跨層重複傳遞造成不一致。
 
 ## 📁 專案結構
 
-```
-ai_controller_pro/
-│
-├── app.py                  # 主程式（後端）
-├── requirements.txt        # Python 套件列表
-├── README.md              # 本說明文件
-│
+```text
+vscode_controller_project3/
+├── app.py                # Flask + Webview 入口
+├── automation.py         # VS Code、自動化流程、螢幕擷取與程式管理
+├── core.py               # 日誌設定、資料模型、JSON 解析工具
+├── services.py           # 配置存取與 Gemini API 客戶端
+├── requirements.txt      # 套件需求
 ├── templates/
-│   └── index.html         # 前端界面
-│
-└── .ai_controller_v2/     # 配置和數據目錄（自動創建）
-    ├── config.json        # 使用者配置
-    ├── screenshots/       # 螢幕截圖
-    └── logs/             # 執行日誌
+│   └── index.html        # 前端介面
+└── ~/.ai_controller_v3/  # 執行時自動建立的資料目錄
+    ├── config.json       # 使用者設定
+    ├── logs/             # 日誌檔案
+    ├── projects/         # 生成專案
+    └── screenshots/      # 螢幕截圖
 ```
 
-## ⚙️ 進階配置
+> 📌 **檔案數量限制**：核心邏輯集中在四個 Python 檔案內，符合「不超過 5 個代碼檔案」的需求。
 
-### 生成參數調整
-- **Temperature (0-2)**：控制輸出的隨機性
-- **Top-P (0-1)**：核心採樣參數
-- **Top-K (1-100)**：候選詞彙數量
-- **Max Tokens**：最大輸出長度
+## 🚀 安裝與啟動
 
-### 自動化設定
-- **自動錯誤修正**：檢測並修正執行錯誤
-- **自動程式碼優化**：改善程式碼品質
-- **自動測試執行**：執行單元測試
-- **監控間隔**：設定檢查頻率
+1. **建立虛擬環境並安裝依賴**
+   ```bash
+   python -m venv venv
+   source venv/bin/activate  # Windows 使用 .\venv\Scripts\activate
+   pip install -r requirements.txt
+   ```
 
-## 🔐 安全性考量
+2. **設定 Gemini API**
+   - 於介面設定頁輸入 API Key，或
+   - 透過 `gcloud auth application-default login` 設定 Google Cloud 憑證。
 
-1. **檔案名稱驗證**
-   - 禁止路徑遍歷攻擊
-   - 過濾特殊字符
+3. **啟動系統**
+   ```bash
+   python app.py
+   ```
+   Flask 伺服器會在背景啟動，並透過 PyWebview 開啟桌面視窗。
 
-2. **程式執行控制**
-   - 超時機制防止無限循環
-   - 進程隔離執行
-   - 資源使用限制
+## 🧭 操作流程
 
-3. **API 金鑰保護**
-   - 本地儲存加密（建議）
-   - 支援環境變數配置
+1. **選擇專案資料夾**：於主控制台點擊「瀏覽」，選定生成程式碼要儲存的位置。
+2. **撰寫 AI 指令**：輸入需求描述，可一併宣告安裝需求（例：`pip install fastapi`）。
+3. **執行自動化**：
+   - 呼叫 Gemini 產出 JSON 化專案結構。
+   - 安裝依賴並寫入檔案。
+   - 自動啟動 VS Code 並開啟核心檔案。
+   - 如主檔案會啟動視窗或網頁，系統會紀錄 PID、開啟網址並擷取畫面。
+4. **擷取截圖**：可指定視窗標題或專案名稱，後端會在完成程式啟動後統一擷取，避免前端重複回填變量。
 
-## 🐛 故障排除
+## 🧱 模組詳解
 
-### 問題：找不到 VS Code
-**解決方案**：確保 `code` 命令已加入 PATH
-```bash
-# 測試命令
-code --version
-```
+### `core.py`
+- 定義日誌設定與執行資料夾。
+- 集中 `AIConfig`、`FileOutput`、`ProjectOutput`、`ProcessResult` 等資料模型。
+- 提供 JSON Schema、系統指令、安裝與檔案儲存工具，確保解析後的資料維持單一來源。
 
-### 問題：螢幕擷取失敗
-**解決方案**：檢查 mss 套件安裝
-```bash
-pip install --upgrade mss
-```
+### `services.py`
+- `ConfigManager`：負責讀寫 `config.json`，寫入時使用 `asdict` 確保結構一致。
+- `GeminiAI`：封裝 API Key 與 Google Cloud Auth 兩種連線方式，並依回應模式切換生成參數。
 
-### 問題：AI 回應格式錯誤
-**解決方案**：確保提示詞包含正確格式指示
-- 程式碼：\`\`\`python ... \`\`\`
-- 檔案名：/*/filename.py/*/
-- 安裝指令：;;;pip install package;;;
+### `automation.py`
+- `VSCodeController`：啟動 VS Code、搜尋視窗與自動開啟多個檔案。
+- `ScreenCapture`：以 PyWinCtl + MSS 擷取指定視窗，統一由後端儲存檔名與路徑。
+- `ProgramManager`：追蹤與終止子行程，並提供 PID、執行狀態與運行時間。
+- `ProcessManager`：整合呼叫 AI、解析、安裝、寫檔、啟動程式、擷取畫面等完整流程，若解析失敗會給出建議與偵錯資訊。
 
-## 📈 未來規劃
+### `app.py`
+- 建立 Flask 路由與 PyWebview 視窗。
+- 轉交請求給服務與自動化模組，確保路由層僅負責資料整合與回應序列化。
 
-- [ ] 支援更多 IDE（IntelliJ IDEA、Sublime Text）
-- [ ] 整合版本控制（Git 自動提交）
-- [ ] 支援多語言程式碼生成
-- [ ] 加入程式碼審查功能
-- [ ] 實現分散式執行
-- [ ] 添加 Web API 介面
-- [ ] 整合 CI/CD 流程
-- [ ] 支援團隊協作功能
+## 🛠 偵錯與日誌
 
-## 🤝 貢獻指南
+- 所有模組共用 `core.py` 初始化的 `logger`，輸出格式包含時間戳與級別。
+- `ProcessManager` 在每個主要步驟皆有 `Step` 紀錄；解析失敗時會回傳「解析錯誤」區塊，包含可能原因與建議。
+- `ProgramManager` 會把執行中的程式以 PID 為 key 儲存在單一定義的字典內，並提供 `/running-programs` API 查詢。
 
-歡迎提交 Pull Request 或開啟 Issue！
+## 🐛 常見問題
 
-1. Fork 本專案
-2. 創建功能分支 (`git checkout -b feature/AmazingFeature`)
-3. 提交更改 (`git commit -m 'Add some AmazingFeature'`)
-4. 推送到分支 (`git push origin feature/AmazingFeature`)
-5. 開啟 Pull Request
+| 問題 | 可能原因 | 建議處理 |
+| --- | --- | --- |
+| 無法找到 VS Code 視窗 | `code` 指令未加入 PATH 或啟動時間過長 | 執行 `code --version` 確認；視情況增加等待時間 |
+| JSON 解析失敗 | 模型輸出非預期格式 | 檢查 `result.output` 內的錯誤說明，改用文本模式或調整提示 |
+| 螢幕截圖為空 | 視窗尚未建立或標題不符 | 於擷取 API 加入正確的視窗標題，或延長等待時間 |
 
-## 📄 授權協議
+## ✅ 測試建議
 
-本專案採用 MIT 授權協議 - 詳見 [LICENSE](LICENSE) 文件
-
-## 🙏 致謝
-
-- Google Gemini AI 團隊
-- PyAutoGUI 和 PyGetWindow 開發者
-- MSS 螢幕擷取庫作者
-- Flask 和 pywebview 社群
-
-## 📞 聯絡方式
-
-- 專案網站：[https://your-project-site.com](https://your-project-site.com)
-- 問題回報：[GitHub Issues](https://github.com/your-repo/issues)
-- 電子郵件：contact@your-domain.com
+1. **API Key 模式**：輸入有效 Key，建立簡單 Python 專案。
+2. **Google Auth 模式**：以 `gcloud` 登入後，確認能生成專案並啟動 VS Code。
+3. **GUI/Web 專案**：測試會開啟視窗的程式，確認自動擷取畫面與 PID 管理功能。
 
 ---
 
-**注意事項**：
-- 本專案為教育和研究目的開發
-- 使用時請遵守相關法律法規
-- AI 生成的程式碼需要人工審查
-- 建議在虛擬環境或容器中執行未經測試的程式碼
-
-**版本歷史**：
-- v2.0.0 (2024-11) - 全面改版，新增螢幕擷取和自動化功能
-- v1.0.0 (2024-10) - 初始版本發布
+如需延伸或整合其他自動化流程，可在既有模組內擴充，並持續遵循「核心變量集中、檔案數量精簡、完整偵錯資訊」的設計原則。
